class ve dosya adları düzeltildi
